Scaffold automated tests
