+++ test/conftest.py
@@ -1,7 +1,7 @@
-import pathlib
+from pathlib import Path
 import pytest
 
 
 @pytest.fixture
 def btl_dir():
-    return pathlib.Path(__file__).parent / "bottlefile_test_data"
+    return Path(__file__).parent / "bottlefile_test_data"
